Bring back useOnboarding.
